requirements.txt
firecrawl-py>=1.0.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.60.0
google-cloud-firestore>=2.14.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
